Equivalent flowObjective implementation from Atrium 2015/A
Implement the flow objective in ODL.
Problem: As the switches have different pipelines, so the application
written for the switch pipeline X can't be used to work with the
switch pipeline Y.

This dependency can be resolved using the flow-objective. Flow
objectives are intent behind the flow installation. The flow installation
objective is translated to flow rules using the Device Drivers.

Currently the there are three Flow Objective: Next, forwarding and
filter.